BlackScholes last version

+++ BlackScholes.h
@@ -2,6 +2,8 @@
 #define BLACKSCHOLES_H
 #include <iostream>
 #include "AbstractOption.h"
+#include "AsianOption.h"
+#include "EuropeanOption.h"
 using namespace std;
 
 class BlackScholes
@@ -11,7 +13,7 @@ public:
 	BlackScholes();
 	
 // Other Methods :
-	double * TimeGeneration(int nbIterations, double init, double end); // h = pas de hachage ou de discrétisation
+	double * TimeGeneration(int nbIterations, int init, int end); // h = pas de hachage ou de discrétisation
 	double GenericGenerationRandomPathsUnderlyingPrices();
 	double GenericComputationOptionPrice();
